Minor fixes to get nav bar working.
